requirements.txt
google-genai
fastapi
uvicorn
